requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
psycopg2-binary==2.9.9
django-cors-headers==4.3.1
Pillow==10.1.0
razorpay==1.4.1
qrcode==7.4.2
python-decouple==3.8
gunicorn==21.2.0
boto3==1.29.7
reportlab==4.0.7
